requirement.txt
asgiref==3.3.1
Django==3.1.3
django-filter==2.4.0
django-redis==4.12.1
django-user-agents==0.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.4.0
importlib-metadata==2.0.0
Markdown==3.3.3
psycopg2-binary==2.8.6
PyJWT==1.7.1
pytz==2020.4
PyYAML==5.3.1
redis==3.5.3
sqlparse==0.4.1
ua-parser==0.10.0
user-agents==2.2.0
zipp==3.4.0
